Rename data to initial state
