+++ src/client/src/containers/AzureFunctionsModal/index.tsx
@@ -11,7 +11,7 @@ import asModal from "../../components/Modal";
 
 import { saveAzureFunctionsSettingsAction } from "../../actions/azureActions/azureFunctionActions";
 import { closeModalAction } from "../../actions/modalActions/modalActions";
-import { azureFunctionModalInitialState } from "../../mockData/cosmosDbModalData";
+import { azureFunctionModalInitialState } from "../../mockData/azureModalInitialStateData";
 import { azureMessages as azureModalMessages } from "../../mockData/azureServiceOptions";
 import { ReactComponent as Spinner } from "../../assets/spinner.svg";
 import { ReactComponent as Cancel } from "../../assets/cancel.svg";
@@ -337,7 +337,6 @@ const AzureFunctionsResourceModal = (props: Props) => {
               className={styles.link}
               href={links[formSectionId]}
               onKeyUp={keyUpHandler}
-
             >
               {rightHeader}
             </a>
