+++ src/client/src/mockData/azureModalInitialStateData.ts
